Login merged in DI
