+++ .idea/misc.xml
@@ -3,13 +3,11 @@
   <component name="DesignSurface">
     <option name="filePathToZoomLevelMap">
       <map>
-        <entry key="../../../../layout/custom_preview.xml" value="0.25" />
-        <entry key="app/src/main/res/drawable/default_state.xml" value="0.25794871794871793" />
+        <entry key="../../../../layout/custom_preview.xml" value="0.1" />
+        <entry key="app/src/main/res/drawable/default_state.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.1" />
         <entry key="app/src/main/res/layout/activity_main2.xml" value="0.1" />
-        <entry key="app/src/main/res/layout/fixture_header.xml" value="0.11574074074074074" />
-        <entry key="app/src/main/res/layout/fixture_item.xml" value="0.375" />
-        <entry key="app/src/main/res/layout/fragment_blank.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/fixture_item.xml" value="0.335" />
         <entry key="app/src/main/res/layout/fragment_dashboard.xml" value="0.3426124197002141" />
         <entry key="app/src/main/res/layout/fragment_draft.xml" value="0.11828774062816616" />
         <entry key="app/src/main/res/layout/fragment_fixtures.xml" value="0.25" />
@@ -19,7 +17,11 @@
         <entry key="app/src/main/res/menu/bottom_nav_menu.xml" value="0.175" />
         <entry key="app/src/main/res/menu/bottom_navigation_menu.xml" value="0.24322916666666666" />
         <entry key="app/src/main/res/menu/options_menu.xml" value="0.24322916666666666" />
+        <entry key="core-ui/src/main/res/drawable/default_state.xml" value="0.1" />
         <entry key="matches-data/src/main/res/font/arvo.xml" value="0.17407407407407408" />
+        <entry key="matches-ui/src/main/res/drawable/ic_baseline_arrow_back_ios_new_24.xml" value="0.22962962962962963" />
+        <entry key="matches-ui/src/main/res/layout/fixture_header.xml" value="0.1" />
+        <entry key="matches-ui/src/main/res/layout/fragment_fixtures.xml" value="0.25" />
       </map>
     </option>
   </component>
