+++ core-ui/src/main/res/drawable/default_state.xml
@@ -6,5 +6,5 @@
         android:width="16dp"
         android:height="16dp"/>
     <solid
-        android:color="@color/lb_grey"/>
+        android:color="@color/cardview_dark_background"/>
 </shape>
