+++ core-ui/src/main/res/drawable/ic_baseline_cancel_24.xml
