+++ core-ui/src/main/res/drawable/ic_check_circle_24.xml
