+++ matches-ui/src/main/res/layout/fragment_fixtures.xml
@@ -35,9 +35,10 @@
             android:layout_height="0dp"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/gameweek_text"
-            app:layout_constraintVertical_bias="0.0" />
+            app:layout_constraintVertical_bias="1.0" />
 
         <ProgressBar
             android:id="@+id/progressBar"
@@ -81,20 +82,24 @@
             android:id="@+id/back_arrow"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:rotationY="-180"
+            android:src="@drawable/ic_baseline_arrow_back_ios_new_24"
             android:visibility="invisible"
+            app:layout_constraintBottom_toBottomOf="@+id/gameweek_text"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-             />
+            tools:visibility="visible" />
 
         <ImageView
             android:id="@+id/next_arrow"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
+            android:rotationY="-180"
+            android:src="@drawable/ic_baseline_arrow_back_ios_new_24"
             android:visibility="invisible"
+            app:layout_constraintBottom_toBottomOf="@+id/gameweek_text"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-             />
+            tools:visibility="visible" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
